pyproject.toml
[project]
name = "littlefish"
version = "1.0.5"
description = "A bot for the `Minesweeper League` app."
authors = [
    {name = "Tony Xiang"},
]
license = {text = "AGPL-3.0"}
requires-python = "<4.0,>=3.9"
dependencies = [
    "nonebot2[fastapi]<3.0.0,>=2.2.0",
    "nonebot-adapter-console<1.0.0,>=0.6.0",
    "nonebot-adapter-onebot<3.0.0,>=2.2.4",
    "nonebot-plugin-apscheduler<1.0.0,>=0.3.0",
    "nonebot-plugin-alconna<1.0.0,>=0.30.0",
    "pycryptodome<4.0.0,>=3.19.1",
    "42points<2.0.0,>=1.2.7",
    "pydantic-settings<3.0.0,>=2.2.0",
]
readme = "README.md"

[tool.black]
line-length = 127
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 127
length_sort = true
skip_gitignore = true
force_sort_within_sections = true

[tool.pylint]
max-line-length = 127
disable = ["invalid-unary-operand-type", "invalid-name", "unspecified-encoding"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"
reportAttributeAccessIssue = false

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
cli = [
    "nb-cli<2.0.0,>=1.2.5",
]
docs = [
    "mkdocs-material<10.0.0,>=9.4.3",
    "mkdocs-git-revision-date-localized-plugin<2.0.0,>=1.2.1",
]

[tool.nonebot]
adapters = [
    { name = "Console", module_name = "nonebot.adapters.console" },
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_alconna"]
plugin_dirs = ["littlefish"]
